PR suggestions made
