+++ app/src/main/res/anim/shake_cycle.xml
@@ -1,4 +1,3 @@
-
 <?xml version="1.0" encoding="utf-8"?>
 <cycleInterpolator xmlns:android="http://schemas.android.com/apk/res/android"
     android:cycles="5" />
