--- app/src/main/res/layout/currency_input_double.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-    <EditText
-        android:id="@+id/primaryValue"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textAlignment="center"
-        android:textSize="24sp"
-        android:textColor="#00000000"
-        android:backgroundTint="@color/purple_200"
-        android:fontFamily="@font/poppins_lightitalic"
-        app:layout_constraintStart_toStartOf="@id/textContainer"
-        app:layout_constraintEnd_toEndOf="@id/textContainer"
-        app:layout_constraintTop_toTopOf="parent"
-        android:layout_marginTop="-10dp"
-        android:inputType="numberDecimal"
-        tools:text="0,00"
-        />
-    <TextView
-        android:id="@+id/textContainer"
-        android:fontFamily="@font/poppins_lightitalic"
-        android:textColor="@color/black"
-        android:layout_marginEnd="5dp"
-        android:layout_marginBottom="10dp"
-        android:textSize="24sp"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:text="0,00"
-        />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
