+++ app/src/main/res/values/dimens.xml
@@ -117,4 +117,9 @@
     <dimen name="recipe_item_width">70dp</dimen>
     <dimen name="item_recipe_border_radius">10dp</dimen>
     <dimen name="item_recipe_icon_size">50dp</dimen>
+
+    <!-- Currency Input -->
+    <dimen name="currency_input_margin_end">5dp</dimen>
+    <dimen name="currency_input_margin_bottom">10dp</dimen>
+    <dimen name="currency_input_margin_top">-10dp</dimen>
 </resources>
